AI_SERVER/requirements.txt
fastapi
uvicorn[standard]
groq
mysql-connector-python
python-dotenv
requests
pydantic